move parser to its own class
